refactor(services): PR-3.6 dataset/ + runtime/ + data_io/ 子包搬迁 — 13 文件
dataset/（顶层 4 + services 2 = 6 文件搬入）：
  studio/datasets.py     → services/dataset/scan.py
  studio/curation.py     → services/dataset/curation.py
  studio/thumb_cache.py  → services/dataset/thumb_cache.py
  studio/browse.py       → services/dataset/browse.py
  services/uploads.py    → services/dataset/uploads.py
  services/tagedit.py    → services/dataset/tagedit.py

runtime/（环境装配 / 自更新 6 文件搬入）：
  services/torch_setup.py            → services/runtime/torch.py
  services/onnxruntime_setup.py      → services/runtime/onnxruntime.py
  services/flash_attention_setup.py  → services/runtime/flash_attention.py
  services/xformers_setup.py         → services/runtime/xformers.py
  services/pending_install.py        → services/runtime/pending_install.py
  services/updater.py                → services/runtime/updater.py

data_io/（1 文件）：
  services/train_io.py → services/data_io/train_io.py

修复发现：
- runtime/torch.py 和 runtime/pending_install.py 内部互引用还在用 *_setup 旧命名（绕过 shim 改 import）
- data_io/train_io.py 含 3 处 inline import 从 services/ 提到 data_io/ 后 from . 路径不对

回归：全测试套 1635 passed（vs 上 PR base 1622，+13 新模块 import smoke 用例）
