+++ studio/services/flash_attention_setup.py
@@ -1,286 +1,10 @@
-"""Flash Attention wheel 查找与安装。
+"""Re-export shim — PR-3 真实模块 studio.services.runtime.flash_attention。
 
-wheel 命名规律（mjun0812/flash-attention-prebuild-wheels）：
-  flash_attn-{fa_ver}+{cuda}{torch}-{pyver}-{pyver}-{platform}.whl
-  例：flash_attn-2.8.3+cu130torch2.11-cp312-cp312-win_amd64.whl
-
-匹配策略：
-- platform：必须精确一致
-- torch：必须精确一致（2.11 ≠ 2.10）
-- CUDA：精确 > 同大版本（cu132 → 接受 cu130，CUDA 小版本向下兼容）
-- Python：必须精确（cp312 wheel 无法在 cp313 上运行，ABI 不同）
-
-公开 API（也是 server 端点 / CLI 用的入口）：
-- `detect_env()` — 当前 Python / CUDA / PyTorch / 平台
-- `current_status()` — flash_attn 是否已装 + 版本
-- `find_candidates(env)` — GitHub releases 列表（带 score / usable / notes）
-- `find_best_wheel(env)` — 最优可用 wheel URL
-- `install(url=None)` — 同步 pip install；url=None → 自动选
+sys.modules 别名让旧路径的 monkeypatch / 私有访问透明转发到真实模块。
+新代码请直接 `from studio.services.runtime.flash_attention import X`。
 """
-from __future__ import annotations
+import sys as _sys
 
-import importlib.metadata
-import json
-import logging
-import platform
-import re
-import subprocess
-import sys
-import urllib.request
-from typing import Any, Optional
+from .runtime import flash_attention as _real
 
-logger = logging.getLogger(__name__)
-
-FA_RELEASES_URL = (
-    "https://api.github.com/repos/mjun0812/flash-attention-prebuild-wheels/releases"
-)
-
-
-def detect_env() -> dict[str, Any]:
-    """检测当前 Python / CUDA / PyTorch / 平台。
-
-    各字段在不可获取时为 None。`platform` 仅返回 `linux_x86_64` / `win_amd64`，
-    其它平台（macOS arm64 / linux aarch64）目前没 prebuilt wheel。
-    """
-    vi = sys.version_info
-    python_tag = f"cp{vi.major}{vi.minor}"
-
-    syst = platform.system().lower()
-    mach = platform.machine().lower()
-    if syst == "linux" and mach == "x86_64":
-        plat = "linux_x86_64"
-    elif syst == "windows" and mach in ("amd64", "x86_64"):
-        plat = "win_amd64"
-    else:
-        plat = None
-
-    # cuda_tag 必须匹配 **PyTorch 编译时**使用的 CUDA runtime —— flash_attn ABI 跟着
-    # torch 走，不是跟着 driver。优先从 `torch.__version__` 的 `+cuXXX` 后缀拿。
-    #
-    # 历史 bug（PR-7）：原实现从 nvidia-smi 拿 cuda_tag。但 `nvidia-smi` 输出的
-    # "CUDA Version: X.Y" 是**驱动支持的最高 CUDA 版本**，与 PyTorch 编译时锁定的 CUDA
-    # 不是一回事。例：5090 driver 报 13.0，venv 装 torch 2.11.0+cu128 → flash_attn 必须
-    # 装 cu128 wheel；原实现却把 cuda_tag 设成 cu130 → 选了 cu130 wheel → ABI 不匹配
-    # → flash_attn import 失败。
-    cuda_tag: Optional[str] = None
-    cuda_ver: Optional[str] = None
-    torch_tag: Optional[str] = None
-    torch_ver: Optional[str] = None
-    try:
-        import torch  # type: ignore[import-not-found]  # noqa: PLC0415
-        torch_ver = torch.__version__
-        v = torch_ver.split("+")[0].split(".")
-        torch_tag = f"torch{v[0]}.{v[1]}"
-        m = re.search(r"\+cu(\d+)", torch_ver)
-        if m:
-            num = m.group(1)
-            cuda_tag = f"cu{num}"
-            # cu128 → 12.8、cu130 → 13.0（最后一位 minor，其余 major）
-            if len(num) >= 2:
-                cuda_ver = f"{num[:-1]}.{num[-1]}"
-    except ImportError:
-        pass
-
-    # nvidia-smi 仍跑一下：torch 没 +cu 后缀（CPU-only build）→ fallback 给 cuda_tag；
-    # driver 版本始终单独存到 `driver_cuda_ver` 供 UI 显示与排错（让用户看到「驱动支持
-    # cu130，PyTorch 是 cu128」这种场景，立刻明白为什么 wheel 应该选 cu128）。
-    driver_cuda_ver: Optional[str] = None
-    try:
-        r = subprocess.run(
-            ["nvidia-smi"], capture_output=True, text=True, timeout=10
-        )
-        if r.returncode == 0:
-            m = re.search(r"CUDA Version:\s*(\d+)\.(\d+)", r.stdout)
-            if m:
-                driver_cuda_ver = f"{m.group(1)}.{m.group(2)}"
-                if cuda_tag is None:
-                    cuda_tag = f"cu{m.group(1)}{m.group(2)}"
-                    cuda_ver = driver_cuda_ver
-    except (subprocess.SubprocessError, OSError):
-        pass
-
-    return {
-        "python_tag": python_tag,
-        "cuda_tag": cuda_tag,
-        "cuda_ver": cuda_ver,
-        "driver_cuda_ver": driver_cuda_ver,
-        "torch_tag": torch_tag,
-        "torch_ver": torch_ver,
-        "platform": plat,
-    }
-
-
-def _parse_wheel(name: str) -> Optional[dict[str, str]]:
-    """从 wheel 文件名解析出 cuda / torch / python / platform 标签。
-
-    例：flash_attn-2.8.3+cu130torch2.11-cp312-cp312-win_amd64.whl
-    → {cuda: "cu130", torch: "torch2.11", python: "cp312", platform: "win_amd64"}
-    """
-    m = re.search(
-        r"\+(cu\d+)(torch[\d.]+)-(cp\d+)-cp\d+-([\w]+)\.whl$", name
-    )
-    if not m:
-        return None
-    return {
-        "cuda": m.group(1),
-        "torch": m.group(2),
-        "python": m.group(3),
-        "platform": m.group(4),
-    }
-
-
-def _cuda_major(tag: str) -> int:
-    """cu130 → 13，cu124 → 12；解析失败返回 -1。"""
-    m = re.search(r"cu(\d+)", tag)
-    return int(m.group(1)) // 10 if m else -1
-
-
-def find_candidates(
-    env: dict[str, Any],
-) -> tuple[list[dict[str, Any]], Optional[str]]:
-    """查询 GitHub Releases，返回 (candidates, fetch_error)。
-
-    candidates 每项：`{url, name, score, notes, usable, tags}`，按 score 降序。
-    `usable=True` 表示当前环境可以直接安装；False 表示 ABI 不兼容（典型 Python 不一致）。
-    `fetch_error` 非 None 表示 GitHub API 请求失败（网络 / 限流 / 解析）；
-    UI 应显示为「无法拉候选列表，可手动粘 URL」。
-    """
-    plat = env.get("platform")
-    torch_tag = env.get("torch_tag")
-    cuda_tag = env.get("cuda_tag")
-    python_tag = env.get("python_tag")
-
-    if not plat:
-        return [], None
-
-    try:
-        req = urllib.request.Request(
-            FA_RELEASES_URL + "?per_page=100",
-            headers={"User-Agent": "AnimaLoraStudio"},
-        )
-        data = json.loads(urllib.request.urlopen(req, timeout=15).read())
-    except Exception as exc:  # noqa: BLE001
-        return [], str(exc)
-
-    # GitHub 频率限制时返回 dict（{"message": "API rate limit exceeded..."}），不是 list
-    if not isinstance(data, list):
-        msg = data.get("message", str(data)) if isinstance(data, dict) else str(data)
-        return [], f"GitHub API 错误: {msg}"
-
-    candidates: list[dict[str, Any]] = []
-    for release in data:
-        for asset in release.get("assets", []):
-            tags = _parse_wheel(asset["name"])
-            if not tags:
-                continue
-            if tags["platform"] != plat:
-                continue
-            if torch_tag and tags["torch"] != torch_tag:
-                continue
-
-            score = 0
-            notes: list[str] = []
-            usable = True
-
-            # Python ABI：严格匹配，不同版本无法使用
-            if python_tag:
-                if tags["python"] == python_tag:
-                    score += 20
-                else:
-                    usable = False
-                    notes.append(
-                        f"Python 不兼容（wheel={tags['python']}，当前={python_tag}）"
-                    )
-
-            # CUDA：同大版本可用，但不如精确匹配
-            if cuda_tag:
-                if tags["cuda"] == cuda_tag:
-                    score += 20
-                elif _cuda_major(tags["cuda"]) == _cuda_major(cuda_tag):
-                    score += 10
-                    notes.append(
-                        f"CUDA 小版本不同（wheel={tags['cuda']}，当前={cuda_tag}，"
-                        f"同大版本应兼容）"
-                    )
-                else:
-                    score -= 5
-                    notes.append(
-                        f"CUDA 大版本不同（wheel={tags['cuda']}，当前={cuda_tag}）"
-                    )
-
-            candidates.append({
-                "url": asset["browser_download_url"],
-                "name": asset["name"],
-                "score": score,
-                "notes": notes,
-                "usable": usable,
-                "tags": tags,
-            })
-
-    return sorted(candidates, key=lambda x: -x["score"]), None
-
-
-def find_best_wheel(env: dict[str, Any]) -> Optional[str]:
-    """返回最优可用 wheel URL；无候选 / 全部 unusable → None。"""
-    candidates, _ = find_candidates(env)
-    for c in candidates:
-        if c["usable"]:
-            return c["url"]
-    return None
-
-
-def current_status() -> dict[str, Any]:
-    """当前 flash_attn 安装状态。"""
-    try:
-        version = importlib.metadata.version("flash_attn")
-        return {"installed": True, "version": version}
-    except importlib.metadata.PackageNotFoundError:
-        return {"installed": False, "version": None}
-
-
-def install(url: Optional[str] = None) -> dict[str, Any]:
-    """安装 flash_attn wheel；url=None 则自动从 GitHub 找最优匹配。
-
-    同步 pip install，可能需要几分钟（远端 wheel ~150MB）。flash_attn 是 C extension，
-    pip 重装后必须重启进程才能切换；返回 `restart_required=True` 让 UI 提示。
-    """
-    env = detect_env()
-
-    if url is None:
-        if not env.get("platform"):
-            raise RuntimeError("不支持的平台（仅 linux_x86_64 / win_amd64）")
-        if not env.get("torch_tag"):
-            raise RuntimeError("未检测到 PyTorch，无法自动匹配 wheel")
-        url = find_best_wheel(env)
-        if not url:
-            raise RuntimeError(
-                f"未找到可用 wheel（Python={env.get('python_tag')}，"
-                f"CUDA={env.get('cuda_tag')}，Torch={env.get('torch_tag')}）。\n"
-                "请在下方候选列表中手动选择，或前往 "
-                "https://github.com/mjun0812/flash-attention-prebuild-wheels/releases 粘贴 URL"
-            )
-
-    r = subprocess.run(
-        [sys.executable, "-m", "pip", "install", url],
-        capture_output=True,
-        text=True,
-    )
-    stdout = r.stdout + r.stderr
-    tail = "\n".join(stdout.splitlines()[-40:])
-
-    if r.returncode != 0:
-        raise RuntimeError(f"pip install 失败:\n{tail}")
-
-    try:
-        importlib.invalidate_caches()
-        version = importlib.metadata.version("flash_attn")
-    except Exception:  # noqa: BLE001
-        version = None
-
-    return {
-        "installed": True,
-        "version": version,
-        "url": url,
-        "stdout_tail": tail,
-        "restart_required": True,
-    }
+_sys.modules[__name__] = _real
