+++ studio/services/uploads.py
@@ -1,122 +1,10 @@
-"""本地上传：单图或 zip 压缩包（自动解压拍平），落盘到 download/。
+"""Re-export shim — PR-3 真实模块 studio.services.dataset.uploads。
 
-用户通过浏览器 file picker / 拖拽上传文件 → server 端解析 → 写入项目的
-`download/` 目录，与 booru 下载共享同一份「全量备份」。
-
-约束：
-- 接受 IMAGE_EXTS 里所有格式的单图（png / jpg / jpeg / webp / bmp / gif）；
-  zip 包内同样按这套白名单提取，原样落盘（不做格式转换）。
-- zip 内的子目录结构会被拍平（取 basename）。
-- 文件名冲突保守跳过，不覆盖、不自动重命名 —— 让用户知道哪里跳过了。
-- zip 损坏时整包跳过并报告，不影响其它文件。
+sys.modules 别名让旧路径的 monkeypatch / 私有访问透明转发到真实模块。
+新代码请直接 `from studio.services.dataset.uploads import X`。
 """
-from __future__ import annotations
+import sys as _sys
 
-import shutil
-import zipfile
-from dataclasses import dataclass, field
-from pathlib import Path
-from typing import BinaryIO, Iterable
+from .dataset import uploads as _real
 
-from ..datasets import IMAGE_EXTS
-
-# PP10 起复用全链路白名单：上传 / 下载 / curation / 训练共用一份。
-ALLOWED_IMAGE_EXTS = IMAGE_EXTS
-ZIP_EXT = ".zip"
-
-
-@dataclass
-class UploadResult:
-    """单次上传调用的汇总结果。"""
-
-    added: list[str] = field(default_factory=list)
-    skipped: list[dict[str, str]] = field(default_factory=list)
-
-    def as_dict(self) -> dict[str, list]:
-        return {"added": self.added, "skipped": self.skipped}
-
-    def merge(self, other: "UploadResult") -> None:
-        self.added.extend(other.added)
-        self.skipped.extend(other.skipped)
-
-
-def _safe_basename(name: str) -> str:
-    """剥掉 zip 内嵌套子目录 / Windows 反斜杠，只留 basename。"""
-    return name.replace("\\", "/").rsplit("/", 1)[-1]
-
-
-def _is_image_ext(name: str) -> bool:
-    return Path(name).suffix.lower() in ALLOWED_IMAGE_EXTS
-
-
-def accept_one(
-    src_name: str, src_stream: BinaryIO, dest_dir: Path
-) -> UploadResult:
-    """处理单个上传文件。
-
-    - IMAGE_EXTS 内任一格式 → 直接拷到 dest_dir（不做格式转换）
-    - zip → 解压所有图片格式（拍平）
-    - 其他 / 没扩展名 → 拒绝
-    """
-    base = _safe_basename(src_name or "")
-    suffix = Path(base).suffix.lower()
-    result = UploadResult()
-    if not base:
-        result.skipped.append({"name": src_name, "reason": "文件名为空"})
-        return result
-
-    dest_dir.mkdir(parents=True, exist_ok=True)
-
-    if suffix in ALLOWED_IMAGE_EXTS:
-        target = dest_dir / base
-        if target.exists():
-            result.skipped.append({"name": base, "reason": "已存在，跳过"})
-        else:
-            with target.open("wb") as fh:
-                shutil.copyfileobj(src_stream, fh)
-            result.added.append(base)
-        return result
-
-    if suffix == ZIP_EXT:
-        try:
-            with zipfile.ZipFile(src_stream) as zf:
-                for info in zf.infolist():
-                    if info.is_dir():
-                        continue
-                    inner_base = _safe_basename(info.filename)
-                    label = f"{base}:{info.filename}"
-                    if not inner_base:
-                        continue
-                    if not _is_image_ext(inner_base):
-                        result.skipped.append(
-                            {"name": label, "reason": "格式不支持"}
-                        )
-                        continue
-                    target = dest_dir / inner_base
-                    if target.exists():
-                        result.skipped.append(
-                            {"name": label, "reason": "已存在，跳过"}
-                        )
-                        continue
-                    with zf.open(info) as src, target.open("wb") as dst:
-                        shutil.copyfileobj(src, dst)
-                    result.added.append(inner_base)
-        except zipfile.BadZipFile:
-            result.skipped.append({"name": base, "reason": "zip 损坏"})
-        return result
-
-    allowed = ", ".join(sorted(ALLOWED_IMAGE_EXTS)) + ", .zip"
-    result.skipped.append(
-        {"name": base, "reason": f"格式不支持（仅 {allowed}）"}
-    )
-    return result
-
-
-def accept_many(
-    files: Iterable[tuple[str, BinaryIO]], dest_dir: Path
-) -> UploadResult:
-    """批量处理；汇总 added / skipped。"""
-    out = UploadResult()
-    for name, stream in files:
-        out.merge(accept_one(name, stream, dest_dir))
-    return out
+_sys.modules[__name__] = _real
